fht-compositor: Layout update
